+++ src/test/java/ru/vyarus/guice/ext/generator/support/classloader/SampleBean.java
@@ -1,0 +1,10 @@
+package ru.vyarus.guice.ext.generator.support.classloader;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 01.06.2016
+ */
+public interface SampleBean {
+
+    String sample();
+}